refactor: extract complete button and variable editor as components
