+++ tasklist/client/src/Tasks/Task/Variables/VariableEditor/styles.module.scss
@@ -14,24 +14,49 @@
  * SUBJECT AS SET OUT BELOW, THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE, AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  * NOTHING IN THIS AGREEMENT EXCLUDES OR RESTRICTS A PARTY’S LIABILITY FOR (A) DEATH OR PERSONAL INJURY CAUSED BY THAT PARTY’S NEGLIGENCE, (B) FRAUD, OR (C) ANY OTHER LIABILITY TO THE EXTENT THAT IT CANNOT BE LAWFULLY EXCLUDED OR RESTRICTED.
  */
-@use '@carbon/layout';
 
-.form {
-  width: 100%;
-  height: 100%;
+
+.list {
+  height: min-content;
 }
 
-.container {
-  padding: 0 0 var(--cds-spacing-05) 0;
+.listCell {
+  vertical-align: middle;
 }
 
-.inlineIcon {
-  margin-right: var(--cds-spacing-01);
+.cellName {
+  width: 200px;
 }
 
-.panelHeader {
+.valueCell {
+  overflow-y: hidden;
+  word-break: break-all;
+  height: min(100px, min-content);
+}
+
+.controlsCell {
+  width: 80px;
+}
+
+.iconButtons {
   display: flex;
-  justify-content: space-between;
-  align-items: center;
-  min-height: layout.to-rem(32px);
+  flex-direction: row;
+  justify-content: flex-end;
+  padding-right: var(--cds-spacing-03);
+
+  &.extraPadding {
+    padding-right: calc(var(--cds-spacing-03) + 32px);
+  }
 }
+
+.scrollableOuter {
+  height: min-content;
+  overflow-y: auto;
+}
+
+.scrollableInner {
+  max-height: 100px;
+  height: min(min-content, 100%);
+  overflow-y: auto;
+}
+
